B5M float -> double + small fixes added debug method for B5M divided alloc and init phases

+++ projects/resources/java/grcuda-benchmark/src/main/java/it/necst/grcuda/benchmark/BenchmarkResults.java
@@ -43,8 +43,8 @@ import java.util.stream.Collectors;
 public class BenchmarkResults {
     List<BenchmarkRecord> phases = new LinkedList<>();
     final BenchmarkConfig config;
-    public float gpu_result; // here we store the result of the gpu computation (if a reduction is done)
-    public float cpu_result; // here we store the result of the cpu computation (if a reduction is done)
+    public double gpu_result; // here we store the result of the gpu computation (if a reduction is done)
+    public double cpu_result; // here we store the result of the cpu computation (if a reduction is done)
     
 
     BenchmarkResults(BenchmarkConfig config){
